Downstream Object, types and Symbol from stage2
As most features have only been implemented in the in-house
linker of the Zig compiler, this commit is a large step to
downstream all those changes back into zwld. Some notable
features are its object parsing, and the usage of string
tables to deduplicate strings. From here, implementing
the remaining features will be easier.