parse and emit producers section for each object
When linking with multiple object files, we first parse all of the
producers sections of each object file and then append the individual
fields to a map. This is to ensure the fields are de-duplicated based
on the version of the field. Finally, we emit all deduplicated fields
into the final binary.